Issue#9 (#10)
The current dynamoDB backend can incur huge costs to the environment during performance tests. This change enables us to reduce that by having an embedded backend in the graph. The two main options are to use an in-memory graph, which does not work well with indices, or an embedded cassandra instance, which does; therefore we will pick the latter.